chore: added kotlin-qa for quality assurance on gradle build files and wartremover for linting scala code.
WartRemover was not available for Scala 3 as a plugin. So, a small extension of the project has been
implemented in the buildSrc subproject providing that functionality.
WartRemover is now configurable in the .wartremover.conf HOCON configuration file.

A new gradle version catalog has been added for the buildSrc subproject, in order to maintain dependencies
differently from the main project. So, in the future it will be easier to transform the buildSrc subproject
into an external dependency.